Add support for a new flag `--rustflag` which passes the Rust compiler flag specified to rustc for the current crate or the current set of crates in a workspace.
Allow setting rustflags for upstream dependencies and/or build scripts but not for the standard library.

Ensure user rustflag entries are sorted.

When adding the lib targets for the current unit also pass down the user specified rustflags for the package. This dependency is typically a special case and the rustflags are required to ensure the unit dependency for the lib target isn't treated as a new unit.

Update the syntax for using the --rustflag feature. Require the `=` delimiter to ensure `clap` doesn't parse a rustflag as a separeate cargo option. Add additional tests for build scripts, proc-macros, multiple targets, workspaces and cross compilation.

Cleanup fingerprint for rustflags and run cargo fmt.